// src/types/api/bindings/Setting.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Column } from "./Column";
import type { Footer } from "./Footer";

export type Setting = { 
/**
 * The ID of the option
 */
id: string, 
/**
 * The name of the option
 */
name: string, 
/**
 * The description of the option
 */
description: string, 
/**
 * View template, used for the title of the embed but can also be used for handling client side errors
 */
view_template: string | null, 
/**
 * Index by
 *
 * If set, all options within the setting will be draggable with the provided index field (must be
 * a integer being set to the position of the item in the list)
 */
index_by: string | null, 
/**
 * The columns for this option
 */
columns: Array<Column>, 
/**
 * The supported operations for this option
 */
operations: Array<string>, 
/**
 * Footer for the setting
 */
footer: Footer | null, 
/**
 * Icon to use for the setting
 */
icon: string | null, 
/**
 * Client side validation script
 */
validation_template: string | null, 
/**
 * Post-send script to run after the setting has been sent to the server
 */
postsend_template: string | null, 
/**
 * DEPRACRATED, but still used in production badgerfang (TO REMOVE)
 */
title_template: string | null, };